Implemented reloading c++ mods
Now copies mods to KHMemoryHook/loaded_mods directory and loads these copies.
On reload it unloads these, deletes them and loads the mods again from the mods directory.
On closing the program a batch script is executed which deletes every mod in the loaded_mods directory after the process finishes, because the accesss is denied on process detaching even though they are unloaded..